refactor(): moving some extended third party libs out to addons folder
