+++ projects/novo-elements/src/addons/ace-editor/ng-package.json
@@ -1,0 +1,5 @@
+{
+  "lib": {
+    "entryFile": "index.ts"
+  }
+}
