+++ projects/novo-elements/src/addons/ckeditor/CKEditor.module.ts
