+++ projects/novo-elements/src/addons/ckeditor/CKEditor.scss
