+++ projects/novo-elements/src/addons/ckeditor/ng-package.json
