+++ projects/novo-elements/src/addons/dragula/Dragula.module.ts
