+++ projects/novo-elements/src/addons/dragula/ng-package.json
