+++ projects/novo-elements/src/addons/index.ts
@@ -1,0 +1,3 @@
+export * from './ace-editor';
+export * from './ckeditor';
+export * from './dragula';
