+++ projects/novo-elements/src/addons/ng-package.json
