+++ projects/novo-elements/src/elements/form/extras/FormExtras.module.ts
@@ -3,9 +3,9 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { NovoPipesModule } from 'novo-elements/pipes';
+import { NovoDragulaModule } from 'novo-elements/addons/dragula';
 import { NovoButtonModule } from 'novo-elements/elements/button';
 import { NovoCheckboxModule } from 'novo-elements/elements/checkbox';
-import { NovoDragulaModule } from 'novo-elements/elements/dragula';
 import { NovoLoadingModule } from 'novo-elements/elements/loading';
 import { NovoPickerModule } from 'novo-elements/elements/picker';
 import { NovoSelectModule } from 'novo-elements/elements/select';
