+++ projects/novo-elements/src/elements/index.ts
@@ -1,4 +1,3 @@
-export * from 'novo-elements/elements/ace-editor';
 export * from 'novo-elements/elements/agenda';
 export * from 'novo-elements/elements/aside';
 export * from 'novo-elements/elements/autocomplete';
@@ -10,14 +9,12 @@ export * from 'novo-elements/elements/card';
 export * from 'novo-elements/elements/category-dropdown';
 export * from 'novo-elements/elements/checkbox';
 export * from 'novo-elements/elements/chips';
-export * from 'novo-elements/elements/ckeditor';
 export * from 'novo-elements/elements/color-picker';
 export * from 'novo-elements/elements/common';
 export * from 'novo-elements/elements/data-table';
 export * from 'novo-elements/elements/date-picker';
 export * from 'novo-elements/elements/date-time-picker';
 export * from 'novo-elements/elements/divider';
-export * from 'novo-elements/elements/dragula';
 export * from 'novo-elements/elements/dropdown';
 export * from 'novo-elements/elements/expansion';
 export * from 'novo-elements/elements/field';
