+++ projects/novo-elements/src/novo-elements.providers.ts
@@ -1,7 +1,7 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
+import { NovoDragulaService } from 'novo-elements/addons';
 import {
   NovoAsideService,
-  NovoDragulaService,
   FieldInteractionApi,
   IMenuOptions, MENU_OPTIONS,
   NovoModalService,
